feat: use `int4` for id columns in db
- auto parse postgres `int4` to `number`

+++ src/db/migrations/1727882828286_create_users_table.ts
@@ -3,8 +3,8 @@ import { type Kysely, sql } from "kysely";
 export async function up(db: Kysely<any>): Promise<void> {
 	await db.schema
 		.createTable("users")
-		.addColumn("id", "bigserial", (col) => col.primaryKey())
-		.addColumn("role_id", "int8", (col) => col.notNull())
+		.addColumn("id", "serial", (col) => col.primaryKey())
+		.addColumn("role_id", "int4", (col) => col.notNull())
 		.addForeignKeyConstraint("role_id_foreign", ["role_id"], "roles", ["id"], (cb) => cb.onDelete("set null"))
 		.addColumn("full_name", "varchar(255)", (col) => col.notNull())
 		.addColumn("email", "varchar(255)", (col) => col.notNull().unique())
